day89-todo_list_web_app/requirements.txt
Flask==2.3.2
Flask-Login
Flask-SQLAlchemy
Werkzeug